CRUD for Adherent

+++ src/main/java/fr/polytech/oeuvres/repositories/OeuvrePretRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OeuvrePretRepository extends JpaRepository<OeuvrePret, Long> {
+public interface OeuvrePretRepository extends JpaRepository<OeuvrePret, Integer> {
 
 }
